tests/wpt/web-platform-tests/css/css-flexbox/parsing/flex-basis-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout: getComputedStyle().flexBasis</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-basis">
<meta name="assert" content="flex-basis computed value is as specified, with length made absolute.">
<meta name="assert" content="flex-basis computed value is non-negative.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("flex-basis", "1px");
test_computed_value("flex-basis", "400%");
test_computed_value("flex-basis", "auto");
test_computed_value("flex-basis", "content");
test_computed_value("flex-basis", "fit-content");
test_computed_value("flex-basis", "min-content");
test_computed_value("flex-basis", "max-content");

test_computed_value("flex-basis", "calc(10px + 0.5em)", "30px");
test_computed_value("flex-basis", "calc(10px - 0.5em)", "0px");
test_computed_value("flex-basis", "calc(10%)", "10%");
// https://github.com/w3c/csswg-drafts/issues/3482
test_computed_value("flex-basis", "calc(0% + 10px)", "calc(0% + 10px)");
test_computed_value("flex-basis", "calc(10% + 0px)", "10%");
</script>
</body>
</html>
